Land V6 live-readiness: single-taker (tennis) real-capital deploy path
Commits the cumulative V6 audit implementation and pre-live hardening that
closes the §5 Final Live-Readiness Gate. Verified green end-to-end: Rust
fmt/clippy + workspace tests, 5/5 cross-language parity, Python ruff/mypy + 331
tests, and capture->normalize->backtest over 48k real captured events.

Risk-gate correctness (Python + Rust parity):
- Side/position-aware cash & exposure so risk-reducing exits are never blocked.
- Split kill switch: realized daily-loss latches; unrealized drawdown is an
  auto-clearing soft-halt that still allows exits.
- Crop model-confidence gate plumbed into the Rust ExternalEdge runtime (was
  Python-only); low-confidence signals now suppressed in both languages.

Deterministic pricing (C3):
- Fixed a Rust pricing scale mismatch (helpers were 1e4 vs the runner's 1e6);
  unified PRICE_SCALE = pricing::PRICE_ONE.
- Edge-preserving cent-flooring of per-side BUY limits in ExternalEdge (Rust)
  and crop/flu/weather (Python), with a half-cent parity case and an AST lint.

Promotion gate & strategy live-readiness:
- verify-strategy enforces real parity coverage, both-language instantiation,
  parity_check, and a no-trade smoke requiring >=1 risk-APPROVED intent.
- flu/crop attach a cached quote snapshot to their orders, fixing silent 100%
  rejection (missing_market_snapshot) and aligning the Python order-snapshot
  gate with the Rust recorded-BBO gate.

Live execution path:
- Startup reconciliation: venue positions/balance seeding + diff report +
  halt-on-breach; idempotent daily-loss restore; Kalshi get_balance/list_positions.
- WS ingest-safety gate, periodic metrics snapshot, file-based kill switch,
  and --live-submit preflight.
- Tennis live sleeve (conservative, operator-review-marked) + updated go-live
  runbook, guarded by a Rust parse test.

Also removes the unused allocator/bus crates.

+++ python/src/eventcontracts/cli/sweep.py
@@ -33,6 +33,7 @@ import tomllib
 from collections.abc import Mapping
 from dataclasses import dataclass
 from datetime import UTC, datetime
+from decimal import Decimal
 from pathlib import Path
 from typing import Any
 
@@ -43,7 +44,7 @@ from eventcontracts.cli.backtest import run_backtest
 from eventcontracts.config import load_sleeve_spec, load_strategy_spec
 from eventcontracts.domain.spec import StrategySpec
 
-ParamValue = str | int | float | bool
+ParamValue = str | int | Decimal | bool
 
 
 def register(subparsers: Any) -> None:
@@ -338,7 +339,7 @@ def _handle(args: argparse.Namespace) -> int:
 
     args.out.parent.mkdir(parents=True, exist_ok=True)
     table = pa.Table.from_pylist(rows, schema=RESULTS_SCHEMA)
-    pq.write_table(table, args.out, compression="snappy")  # type: ignore[no-untyped-call]
+    pq.write_table(table, args.out, compression="snappy")
     error_count = sum(1 for r in rows if r["error"])
     print(
         f"sweep: wrote {len(rows)} rows to {args.out}"
